requirements.txt
Quart==0.19.6
jsonschema==4.23.0
#grpc
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.3
python-dotenv==1.0.1
requests==2.32.3
#pandas==2.2.3
#scikit-learn==1.5.2